service configuration system is now modular, better organized
